// utils/types/board.types.ts
export interface Comment {
  id: string;
  content: string;
  createdAt: string;
}

export interface Card {
  id: string;
  title: string;
  comments: Comment[];
}

export interface List {
  id: string;
  title: string;
  cards: Card[];
}

export interface Board {
  id: string;
  title: string;
  lists: List[];
}